pyproject.toml
[tool.poetry]
name = "website_finder_refinement"
version = "0.1.0"
description = "website_finder_refinement using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^0.51.1"}
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
aiohttp = "^3.9.5"
langchain = "0.2.14"
nest-asyncio = "^1.6.0"
json-repair = "^0.25.3"
validators = "^0.33.0"
asyncio = "^3.4.3"
pytest-playwright = "^0.5.1"
tldextract = "5.1.2"
scrapegraphai = "1.16.0"
streamlit = "^1.38.0"
watchdog = "^4.0.2"

[tool.poetry.scripts]
website_finder_refinement = "website_finder_refinement.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"